backend/requirements.txt
fastapi==0.115.7
uvicorn[standard]==0.34.0
sqlalchemy==2.0.38
psycopg[binary]==3.2.4
pydantic==2.10.6
pydantic-settings==2.7.1
python-multipart==0.0.20
passlib==1.7.4
python-jose==3.3.0
celery==5.4.0
redis==5.2.1
feedparser==6.0.11
pandas==2.2.3
scikit-learn==1.6.1
numpy==2.2.2
joblib==1.4.2
pgvector==0.3.6
requests==2.32.3
prometheus-client==0.21.1
pytest==8.3.4
